ds - refactored controllers
