+++ src/main/java/hello/controllers/EarthquakeController.java
@@ -1,4 +1,4 @@
-package hello;
+package hello.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
@@ -13,44 +13,15 @@ import java.util.HashMap;
 
 import com.nimbusds.oauth2.sdk.client.ClientReadRequest;
 import hello.geojson.FeatureCollection;
+import hello.EarthquakeQueryService;
+import hello.EqSearch;
 
 
 @Controller
-public class WebController {
+public class EarthquakeController {
 
     @Autowired
     private ClientRegistrationRepository clientRegistrationRepository;
-
-    @GetMapping("/")
-    public String getHomepage(Model model, OAuth2AuthenticationToken oAuth2AuthenticationToken) {
-        return "index";
-    }
-
-    @GetMapping("/login")
-    public String getLoginPage(Model model, OAuth2AuthenticationToken oAuth2AuthenticationToken) {
-
-        Map<String, String> urls = new HashMap<>();
-
-        // get around an unfortunate limitation of the API
-        @SuppressWarnings("unchecked") Iterable<ClientRegistration> iterable = ((Iterable<ClientRegistration>) clientRegistrationRepository);
-        iterable.forEach(clientRegistration -> urls.put(clientRegistration.getClientName(),
-                "/oauth2/authorization/" + clientRegistration.getRegistrationId()));
-
-        model.addAttribute("urls", urls);
-        return "login";
-    }
-
-    @GetMapping("/page1")
-    public String getPage1(Model model, OAuth2AuthenticationToken oAuth2AuthenticationToken) {
-
-        return "page1";
-    }
-
-    @GetMapping("/page2")
-    public String getPage2(Model model, OAuth2AuthenticationToken oAuth2AuthenticationToken) {
-
-        return "page2";
-    }
 
     @GetMapping("/earthquakes/search")
     public String getEarthquakesSearch(Model model, OAuth2AuthenticationToken oAuth2AuthenticationToken,
